Separate production code and tests
