+++ README.md
@@ -12,11 +12,11 @@ store with a certificate and matching private key inside.
 
 ### Points of interest
 
-* [Key "repository"](test/index.test.js#L8)
-* [Secret "repository"](test/index.test.js#L37)
-* [Certificate "repository"](test/index.test.js#L58)
-* [Empty credentials for connecting to Lowkey Vault](test/index.test.js#L70)
-* [Tests](test/index.test.js#L80)
+* [Key "repository"](test/index.js#L5)
+* [Secret "repository"](test/index.js#L34)
+* [Certificate "repository"](test/index.js#L55)
+* [Empty credentials for connecting to Lowkey Vault](test/index.test.js#L11)
+* [Tests](test/index.test.js#L21)
 
 ### Usage
